% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_libraries_EI.R
\name{assign_ri}
\alias{assign_ri}
\title{Assign experimental RI to compounds in the msp file}
\usage{
assign_ri(lib, ri_table, polarity = "semi-polar")
}
\arguments{
\item{lib}{The EI library generated by \code{read_lib}.}

\item{ri_table}{The RI table cleaned up by \code{extract_ri}.}

\item{polarity}{The polarity of the column. Can be either "semi-polar",
"non-polar", or "polar".}
}
\value{
A \code{list} with experimental RI assigned.
}
\description{
\code{assign_ri} offers a way to assign experimental RI to the msp file if
you have NIST library installed.
}
\details{
Depending on the column polarity, experimental RI can be assigned to
compounds in the msp file. Providing that "capillary" GC columns are
commonly used. This function will only keep RI records from "capillary"
columns and "Lee RI" will be removed. When there are multiple records for
a single compound, the median RI will be used and if the standard deviation
is higher than 30, this value will be discarded. This function supports
parallel computing.
}
