% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_libraries_EI_and_MS2.R
\name{assign_smiles}
\alias{assign_smiles}
\title{Assign SMILES}
\usage{
assign_smiles(file, structure_data, match = "name")
}
\arguments{
\item{file}{The library generated by \code{read_lib}.}

\item{structure_data}{the correspondent structure data generated by
\code{extract_structure}.}

\item{match}{Correspondence can be done by either "name" or "inchikey".}
}
\value{
A \code{list} with SMILES assigned.
}
\description{
\code{assign_smiles} offers a way to assign SMILES to the library obtained
from NIST format (exported by Lib2NIST with structure information separately
stored in mol files).
}
\details{
The msp file obtained from Lib2NIST has no SMILES and the structure
information is stored in multiple mol files. After transforming all mol files
into a single sdf file by \code{combine_mol2sdf} and retrieving structure
information by \code{extract_structure}, SMILES is available. This function
provides a way to assign SMILES to correspondent compound in the msp file.
If you are working with Linux-based or Mac OS system, it is better to use
"inchikey" for matching. However, if you are working with Windows, the only
option is to use "name" for matching, which is a kind of compromise as
some chemicals in the *.mol files do not have full chemical names. Hence they
will not be matched. This function is useful for both EI and MS2 libraries.
This function supports parallel computing.
}
