% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_libraries_MS2.R
\name{change_meta}
\alias{change_meta}
\title{Change meta data}
\usage{
change_meta(file, CE = NA, instrument = NA, comment = NA, add = FALSE)
}
\arguments{
\item{file}{The in-house library generated by \code{read_lib}.}

\item{CE}{User defined collision energy. If no CE is supplied, the
CollisionEnergy field will not be changed.}

\item{instrument}{User define instrument type. If no instrument is supplied,
the InstrumentType field will not be changed.}

\item{comment}{User define comment, e.g., Principle investigator, data
collector, laboratory, etc.If no comment is supplied, the Comment field
will not be changed. If you want to add new comment, please set "add = TRUE",
then old and new comment will be separated by ";". Otherwise, the old
comment will be covered by the new one.}

\item{add}{A logical scalar. Whether to keep the old comment and add new
comment behind or just replace the old comment. TRUE or FALSE.}
}
\value{
A \code{list} with meta data assigned.
}
\description{
\code{change_meta} offers a way to change meta data (mainly used for in-house
library).
}
\details{
When you build your own mass spectral library (either EI or MS2 library),
you might want to add or change some meta data, such as collision energy,
instrument,and comment. This function provides an easy way to achieve this.
}
