% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_libraries_EI.R
\name{clean_ri_dat}
\alias{clean_ri_dat}
\title{Extract experimental RI from NIST}
\usage{
clean_ri_dat(file)
}
\arguments{
\item{file}{The "ri.dat" file in the installation path
(e.g., "~/Programs/nist17/mssearch").}
}
\value{
A cleaned data.frame containing experimental RI from NIST
}
\description{
\code{clean_ri_dat}, an interanl function, offers a way to extract RI from
the "ri.dat" file.
}
\details{
Once you have NIST library installed, there will be a "ri.dat" file in the
installation path (e.g., "~/Programs/nist17/mssearch"). This file
contains all experimental RI in the NIST library but it is not human readable.
This function provides a way to convert the "ri.dat" file into a data.frame,
so that we can better leverage the RI information present in the NIST library
and to incorporate them into the msp file.
}
