% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_libraries_EI.R
\name{clean_user_dbu}
\alias{clean_user_dbu}
\title{Extract InChIKey for compounds that have experimental RI}
\usage{
clean_user_dbu(file)
}
\arguments{
\item{file}{The "USER.DBU" file in the installation path
(e.g., "~/Programs/nist17/mssearch")}
}
\value{
A data.frame containing four variables, Name, InChIKey, ID,
and "Formula"
}
\description{
\code{clean_user_dbu}, an internal function, offers a way to extract InChIKey
for compounds that have experimental RI from the "USER.DBU" file.
}
\details{
RI values in the cleaned RI table obtained by \code{\link{clean_ri_dat}}
cannot be linked to compounds in the msp file. Providing that the "USER.DBU"
file in the installation path (e.g., "~/Programs/nist17/mssearch") contains
InChIKey of each compound in RI table. However this file is not human
readable. Therefore, this function provides a way to clean the "USER.DBU"
file. Then, we can link experimental RI values to the compounds in the msp
file.
}
