% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_libraries_MS2.R
\name{complete_gnps}
\alias{complete_gnps}
\title{Add formula to the gnps library and remove wrong SMILES}
\usage{
complete_gnps(lib)
}
\arguments{
\item{lib}{The \code{list} generated by \code{read_lib} from mgf library.}
}
\value{
A \code{list} with molecular formula assigned
}
\description{
\code{complete_gnps} offers a way to complete the molecular formula filed in
the mgf file.
}
\details{
The mgf file downloaded from GNPS has no molecular formula (MF). Therefore, this
function tries to calculate the MF from the SMILES (if it exists).
}
