% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_libraries_MS2.R
\name{complete_mgf}
\alias{complete_mgf}
\title{Add formula to the mgf library}
\usage{
complete_mgf(file)
}
\arguments{
\item{file}{The \code{list} generated by \code{read_lib} from mgf library.}
}
\value{
A \code{list} with molecular formula assigned
}
\description{
\code{complete_mgf} offers a way to complete the molecular formula filed in
the mgf file.
}
\details{
The mgf file downloaded from GNPS has no molecular formula (MF). Therefore, this
function tries to calculate the MF from the SMILES (if it exists).
}
