% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EI}
\alias{EI}
\title{An example \code{list} read from EI mass spectral library}
\format{
An object of class \code{list} of length 4.
}
\usage{
EI
}
\description{
An example \code{list} read from EI mass spectral library
}
\keyword{datasets}
