% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_libraries_EI.R
\name{extract_ri}
\alias{extract_ri}
\title{Extract experimental RI from NIST library}
\usage{
extract_ri(ri_dat, user_dbu)
}
\arguments{
\item{ri_dat}{The "ri.dat" file in the installation path
(e.g., "~/Programs/nist17/mssearch/nist_ri").}

\item{user_dbu}{The "USER.DBU" file in the installation path
(e.g., "~/Programs/nist17/mssearch/nist_ri")}
}
\value{
A \code{data.frame} containing experimental RI and InChIKey assigned.
}
\description{
\code{extract_ri} offers a way to extract experimental RI from the NIST
library if you have it installed.
}
\details{
Once you have NIST library installed, there will be a "ri.dat" file in the
installation path (e.g., "~/Programs/nist17/mssearch"). This file
contains all experimental RI in the NIST library but it is not human readable.
This function firstly convert the "ri.dat" file into a data.frame. However,
it is tricky to link RI values in the cleaned RI table to compounds in the
msp file. Providing that the "USER.DBU" file in the installation path
(e.g., "~/Programs/nist17/mssearch") contains InChIKey of each compound in
RI table, we can assign correspondent InChIKey to the RI table, but this
file is not human readable. Therefore, this function secondly provides a way
to clean the "USER.DBU" file and then assign correspondent InChIKey to the
RI table.
}
