% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MS2_mgf}
\alias{MS2_mgf}
\title{An example \code{list} read from MS2 mass spectral library in mgf format}
\format{
An object of class \code{list} of length 4.
}
\usage{
MS2_mgf
}
\description{
An example \code{list} read from MS2 mass spectral library in mgf format
}
\keyword{datasets}
