% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MS2_msp}
\alias{MS2_msp}
\title{An example \code{list} read from MS2 mass spectral library in msp format}
\format{
An object of class \code{list} of length 4.
}
\usage{
MS2_msp
}
\description{
An example \code{list} read from MS2 mass spectral library in msp format
}
\keyword{datasets}
