% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_libraries.R
\name{read_multilibs}
\alias{read_multilibs}
\title{A wrapper to read multiple msp files at a time}
\usage{
read_multilibs(folder)
}
\arguments{
\item{folder}{The folder that contains multiple msp files.}
}
\value{
A single \code{list} combining all msp files
}
\description{
\code{read_multilibs} offers a way to read multiple msp files at a time and
combine them into a single file.
}
\details{
When you are building your in-house libraries, you may probably have multiple
msp files at hand (e.g., one msp for one group of compounds). To avoid
empolying \code{read_lib} several times, this function provides a way to read
all these files at once.
}
