% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_libraries_EI.R
\name{remove_ri}
\alias{remove_ri}
\title{Remove retention index for EI libraries}
\usage{
remove_ri(lib)
}
\arguments{
\item{lib}{The \code{list} generated by \code{read_lib}.}
}
\value{
A \code{list} without RI
}
\description{
\code{remove_ri} offers a way to remove all RI for EI libraries.
}
\details{
This function supports parallel computing.
}
