% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_libraries_MS2.R
\name{remove_rt}
\alias{remove_rt}
\title{Remove retention time for MS2 libraries}
\usage{
remove_rt(file)
}
\arguments{
\item{file}{The \code{list} generated by \code{read_lib}.}
}
\value{
A \code{list} without retention time
}
\description{
\code{remove_rt} offers a way to remove all retention time for MS2 library.
}
\details{
This function supports parallel computing.
}
