% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_libraries_EI_and_MS2.R
\name{reorganize_mona}
\alias{reorganize_mona}
\title{Reorganize MoNA library}
\usage{
reorganize_mona(lib)
}
\arguments{
\item{lib}{The MoNA library generated by \code{read_lib}.}
}
\value{
A \code{List} with SMILES retrieved.
}
\description{
\code{reorganize_mona} offers a way to reorganize MoNA libray,
mainly to retrieve SMILES from the "Comments" field.
}
\details{
The msp file from MoNA has no "SMILES" field but has SMILES information
stored in the "Comments" field. Therefore, This function tries to retrieve
SMILES from the "Comments" field. This function supports parallel computing.
}
