% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_msp.R
\name{write_MS2_msp}
\alias{write_MS2_msp}
\title{Write MS2 library into a msp file}
\usage{
write_MS2_msp(lib, filename)
}
\arguments{
\item{lib}{The organized MS2 library}

\item{filename}{The location and filename of the msp file to be exported,
e.g., "/home/exported.msp".}
}
\value{
No return but create a msp file
}
\description{
\code{write_EI_msp} offers a way to write the organized  MS2 library into a
msp file.
}
