% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_libraries.R
\name{read_lib}
\alias{read_lib}
\title{Read msp/mgf mass spectral libraries}
\usage{
read_lib(file, format = "msp", type = "MS2", remove_ri = TRUE)
}
\arguments{
\item{file}{Mass spectral library in msp or mgf format.}

\item{format}{The format of the library, either msp or mgf.}

\item{type}{The type of the library, either EI or MS2.}

\item{remove_ri}{A logical scalar only used in case of EI mass spectral
library. Should retention index (RI) be removed?}
}
\value{
A \code{list} with each spectral entry as a list element for further
processing.
}
\description{
\code{read_lib} read mass spectral libraries into R for further processing.
}
\details{
This is a generic function to read either EI or MS2 mass spectral libraries.
The library can be either in msp or mgf form. For this reason, it is required
to set the format and the type of the input library. The default is MS2 in
msp format. In the case of EI mass spectral library, an additional boolean
parameter "remove_ri" can be set to remove or keep the retention index (RI).
This function supports parallel computing making use of the \pkg{future.apply}.
Please see the vignette for more details.
}
\examples{
EI <- system.file("EI.msp", package = "mspcompiler")
MS2_mgf <- system.file("MS2.mgf", package = "mspcompiler")

EI <- read_lib(EI, format = "msp", type = "EI", remove_ri = FALSE)
MS2_mgf <- read_lib(MS2_mgf, format = "mgf", type = "MS2")
}
